+++ services/transaction-scheduler/store/postgres/migrations/01_create_tables.go
@@ -56,9 +56,8 @@ CREATE TABLE jobs (
     type TEXT NOT NULL,
     transaction_id INTEGER NOT NULL REFERENCES transactions(id),
 	labels jsonb,
-	internal_data jsonb,
+	annotations jsonb,
 	created_at TIMESTAMPTZ DEFAULT (now() at time zone 'utc') NOT NULL,
-	updated_at TIMESTAMPTZ DEFAULT (now() at time zone 'utc') NOT NULL,
 	UNIQUE(uuid)
 );
 
@@ -71,19 +70,6 @@ CREATE TABLE logs (
 	created_at TIMESTAMPTZ DEFAULT (now() at time zone 'utc') NOT NULL,
 	UNIQUE(uuid)
 );
-
-CREATE OR REPLACE FUNCTION updated() RETURNS TRIGGER AS 
-	$$
-	BEGIN
-		NEW.updated_at = (now() at time zone 'utc');
-		RETURN NEW;
-	END;
-	$$ LANGUAGE plpgsql;
-
-CREATE TRIGGER job_trigger
-	BEFORE UPDATE ON jobs
-	FOR EACH ROW 
-	EXECUTE PROCEDURE updated();
 `)
 	if err != nil {
 		log.WithError(err).Error("Could not create tables")
@@ -97,15 +83,11 @@ CREATE TRIGGER job_trigger
 func dropContextTable(db migrations.DB) error {
 	log.Debug("Dropping tables")
 	_, err := db.Exec(`
-DROP TRIGGER job_trigger ON jobs;
-
 DROP TABLE logs;
 DROP TABLE jobs;
 DROP TABLE transaction_requests;
 DROP TABLE schedules;
 DROP TABLE transactions;
-
-DROP FUNCTION updated();
 `)
 	if err != nil {
 		log.WithError(err).Error("Could not drop tables")